core: Reimplement Document.recover_chunks to get rid of Separator chunks
The initial solution was elegant, but it runs into an issue in Lean.  Given
these chunks:

  - "#check nat"
  - "#check 1"

We produced these:

  - "#check nat"
  - Separator("\n")
  - "#check 1"

But then the answer to "#check nat" returned by Lean covered that *and the
newline*; and then when rechunking the first like would be considered a partial
sentence with no output, and the second line (the separator) would be considered
the real end of the sentence and hence get the message, which would be lost once
separators were removed.

This solution just adds the separator at the end of each fragment instead.