fixup issues with #check
previously, `#check`s on their own (and similar) didn't work. this is because the file got parsed too quickly, and we didn't wait for all the file to be processed before returning (as the AST didn't pick up any points to check at). this forces a wait until the `current_tasks` are empty, which fixes this error by default.